requirements.txt
streamlit==1.45.1
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.19
openai==1.82.1
sentence-transformers==4.1.0
chromadb==1.0.12
requests==2.32.3
python-dotenv==1.1.0
datasets==3.6.0